LayoutTests/media/video-display-aspect-ratio.html

<video controls></video>
<script src=../http/tests/media/media-file.js></script>
<script src=../http/tests/media/video-test.js></script>
<script>
    testExpected("video.videoWidth", 0, "==");
    testExpected("video.videoHeight", 0, "==");

    waitForEvent('loadedmetadata', function () {
        testExpected("video.videoWidth", 426, "==");
        testExpected("video.videoHeight", 240, "==");
        endTest();
    } );

    video.src = findMediaFile("video", "content/test-par-16-9");
</script>
